Safari-609.1.11.1/LayoutTests/fast/events/ios/rotation/zz-no-rotation-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS window.innerWidth is 320
PASS window.innerHeight is 548
This test checks that the view is not left in a rotated state after the previous tests.
